+++ packages/core/base/src/index.ts
@@ -1,1 +1,1 @@
-export * from './utils.js';
+export * from './helloWorld';
